+++ pkg/cache/badger_test.go
