+++ pkg/server/session.go
